requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
transformers==4.33.0
torch==2.0.1
faiss-cpu==1.7.4
plotly==5.17.0